[#24] Cosmetic changes
